feat: 이벤트 재오픈 request 생성
