+++ src/main/java/org/devcourse/resumeme/business/event/controller/dto/EventUpdateRequest.java
@@ -1,5 +1,7 @@
 package org.devcourse.resumeme.business.event.controller.dto;
 
+import org.devcourse.resumeme.business.event.service.vo.EventReopen;
+import org.devcourse.resumeme.business.event.service.vo.EventInfoUpdate;
 import org.devcourse.resumeme.business.event.service.vo.EventUpdateVo;
 import org.devcourse.resumeme.common.domain.Position;
 
@@ -11,12 +13,19 @@ import static org.devcourse.resumeme.common.util.Validator.notNull;
 public record EventUpdateRequest(EventInfoRequest info, EventTimeRequest time, List<String> positions) {
 
     public EventUpdateVo toVo(Long eventId) {
+        if (info == null && time == null && positions == null) {
+            return new EventReopen(eventId);
+        }
+
+        notNull(info);
+        notNull(time);
         notNull(positions);
+
         List<Position> positionEntities = positions.stream()
                 .map(position -> Position.valueOf(position.toUpperCase()))
                 .toList();
 
-        return new EventUpdateVo(eventId, info.title, info.content, info().maximumAttendee, positionEntities,
+        return new EventInfoUpdate(eventId, info.title, info.content, info.maximumAttendee, positionEntities,
                 time.openDateTime, time.closeDateTime, time.endDate);
     }
